Move new PathAttribute and friends into bgp::path_attributes
