+++ src/bgp/message/update_builder.rs
@@ -1,1982 +1,801 @@
-use std::net::Ipv6Addr;
+use std::convert::Infallible;
+use std::fmt;
+use std::iter::Peekable;
+use std::net::{IpAddr, Ipv6Addr};
 
-use octseq::OctetsBuilder;
+use bytes::BytesMut;
+use octseq::{FreezeBuilder, Octets, OctetsBuilder, ShortBuf};
 
+use crate::bgp::aspath::AsPath;
 use crate::bgp::communities::StandardCommunity;
+use crate::bgp::message::{Header, MsgType, SessionConfig, UpdateMessage};
+use crate::bgp::message::attr_change_set::AttrChangeSet;
 use crate::bgp::message::nlri::Nlri;
 use crate::bgp::message::update::{AFI, SAFI, NextHop};
+use crate::bgp::types::OriginType;
+use crate::util::parser::ParseError;
 
-use super::update::ComposeError;
+//use rotonda_fsm::bgp::session::AgreedConfig;
 
-// just drafting ideas, not used right now
-#[allow(dead_code)]
-pub mod new_pas {
+use crate::bgp::path_attributes as new_pas;
+use new_pas::PathAttributeType;
+use new_pas::Attribute; // trait
 
-    use std::fmt;
-    use std::net::Ipv4Addr;
+//------------ UpdateBuilder -------------------------------------------------
+#[derive(Debug)]
+pub struct UpdateBuilder<Target> {
+    target: Target,
+    //config: AgreedConfig, //FIXME this lives in rotonda_fsm, but that
+    //depends on routecore. Sounds like a depedency hell.
+    announcements: Vec<Nlri<Vec<u8>>>,
+    announcements_len: usize,
+    withdrawals: Vec<Nlri<Vec<u8>>>,
+    withdrawals_len: usize,
+    addpath_enabled: Option<bool>, // for conventional nlri (unicast v4)
+    //attributes: Vec<PathAttribute<'a, Vec<u8>>>, // XXX this lifetime is..
+                                                    //not nice
+    // attributes:
+    origin: Option<new_pas::Origin>,
+    aspath: Option<AsPath<Vec<u8>>>,
+    nexthop: Option<new_pas::NextHop>,
+    multi_exit_disc: Option<new_pas::MultiExitDisc>,
+    local_pref: Option<new_pas::LocalPref>,
 
-    use log::debug;
-    use octseq::{Octets, OctetsBuilder, Parser};
+    standard_communities_builder: Option<StandardCommunitiesBuilder>,
 
-    use crate::asn::Asn;
-    use crate::bgp::aspath::HopPath;
-    use crate::bgp::message::update_builder::{
-        MpReachNlriBuilder,
-        MpUnreachNlriBuilder
-    };
-    use crate::bgp::message::SessionConfig;
-    use crate::bgp::types::{AFI, SAFI};
-    use crate::util::parser::{ParseError, parse_ipv4addr};
+    attributes_len: usize,
+    total_pdu_len: usize,
 
-    use super::ComposeError;
+    // MP_REACH_NLRI and MP_UNREACH_NLRI can only occur once (like any path
+    // attribute), and can carry only a single tuple of (AFI, SAFI).
+    // My interpretation of RFC4760 means one can mix conventional
+    // NLRI/withdrawals (so, v4 unicast) with one other (AFI, SAFI) in an
+    // MP_(UN)REACH_NLRI path attribute.
+    // Question is: can one also put (v4, unicast) in an MP_* attribute, and,
+    // then also in the conventional part (at the end of the PDU)? 
+    mp_reach_nlri_builder: Option<MpReachNlriBuilder>,
+    mp_unreach_nlri_builder: Option<MpUnreachNlriBuilder>,
+}
 
-    #[derive(Copy, Clone, Debug, Eq, PartialEq)]
-    pub struct Flags(u8);
+impl<Target: OctetsBuilder> UpdateBuilder<Target> {
 
-    impl Flags {
-        // 0 1 2 3 4 5 6 7
-        //
-        // 0: optional (1 == optional)
-        // 1: transitive (1 == transitive) (well-known attr are transitive)
-        // 2: partial 
-        // 3: extended length (0 -> 1 byte length, 1 -> 2 byte length)
-        // 4-7: MUST be 0 when sent, ignored when received
-        const OPT_NON_TRANS: u8 = 0b1000_0000;
-        const OPT_TRANS: u8     = 0b1100_0000;
-        const WELLKNOWN: u8     = 0b0100_0000;
+    const MAX_PDU: usize = 4096; // XXX should come from NegotiatedConfig
 
-        const EXTENDED_LEN: u8  = 0b0001_0000;
-        const PARTIAL: u8       = 0b0010_0000;
+    pub fn from_target(mut target: Target) -> Result<Self, ShortBuf> {
+        //target.truncate(0);
+        let mut h = Header::<&[u8]>::new();
+        h.set_length(19 + 2 + 2);
+        h.set_type(MsgType::Update);
+        let _ =target.append_slice(h.as_ref());
 
-        /// Returns true if the optional flag is set.
-        pub fn is_optional(self) -> bool {
-            self.0 & 0x80 == 0x80
-        }
+        Ok(UpdateBuilder {
+            target,
+            announcements: Vec::new(),
+            announcements_len: 0,
+            withdrawals: Vec::new(),
+            withdrawals_len: 0,
+            addpath_enabled: None,
 
-        /// Returns true if the transitive bit is set.
-        pub fn is_transitive(self) -> bool {
-            self.0 & 0x40 == 0x40
-        }
+            //attributes:
+            origin: None,
+            aspath: None,
+            nexthop: None,
+            multi_exit_disc: None,
+            local_pref: None,
 
-        /// Returns true if the partial flag is set.
-        pub fn is_partial(self) -> bool {
-            self.0 & 0x20 == 0x20
-        }
+            standard_communities_builder: None,
 
-        /// Returns true if the extended length flag is set.
-        pub fn is_extended_length(self) -> bool {
-            self.0 & 0x10 == 0x10
-        }
-
+            attributes_len: 0,
+            total_pdu_len: 19 + 2 + 2,
+            mp_reach_nlri_builder: None,
+            mp_unreach_nlri_builder: None,
+        })
     }
 
-    impl From<u8> for Flags {
-        fn from(u: u8) -> Flags {
-            Flags(u)
-        }
-    }
+    //--- Withdrawals
 
-    impl From<Flags> for u8 {
-        fn from(f: Flags) -> u8 {
-            f.0
-        }
-    }
-
-
-    pub trait AttributeHeader {
-        const FLAGS: u8;
-        const TYPECODE: u8;
-    }
-
-    macro_rules! attribute {
-        ($name:ident($data:ty),
-         $flags:expr,
-         $typecode:expr
-         ) => {
-
-            #[derive(Debug, Eq, PartialEq)]
-            pub struct $name($data);
-            impl $name {
-                pub fn new(data: $data) -> $name {
-                    $name(data)
-                }
-
-                pub fn inner(self) -> $data {
-                    self.0
-                }
-            }
-
-            impl std::convert::AsRef<$data> for $name {
-                fn as_ref(&self) -> &$data {
-                    &self.0
-                }
-            }
-
-            impl std::convert::AsMut<$data> for $name {
-                fn as_mut(&mut self) -> &mut $data {
-                    &mut self.0
-                }
-            }
-
-
-            /*
-            impl std::ops::Deref for $name {
-                type Target = $data;
-
-                fn deref(&self) -> &Self::Target {
-                    &self.0
-                }
-            }
-
-            impl std::ops::DerefMut for $name {
-                fn deref_mut(&mut self) -> &mut Self::Target {
-                    &mut self.0
-                }
-            }
-            */
-
-            impl AttributeHeader for $name {
-                const FLAGS: u8 = $flags;
-                const TYPECODE: u8 = $typecode;
-            }
-        }
-    }
-
-    macro_rules! path_attributes {
-        (
-            $(
-                $typecode:expr => $name:ident($data:ty), $flags:expr
-            ),+ $(,)*
-        ) => {
-
-            #[derive(Debug, Eq, PartialEq)]
-            pub enum PathAttribute {
-                $( $name($name) ),+,
-                Unimplemented(UnimplementedPathAttribute),
-                Invalid(Flags, u8, Vec<u8>),
-            }
-
-            impl PathAttribute {
-                pub fn compose<Target: OctetsBuilder>(
-                    &self,
-                    target: &mut Target
-                ) -> Result<(), ComposeError> {
-
-                    match self {
-                    $(
-                        PathAttribute::$name(i) => {
-                            i.compose(target)
-                                .map_err(|_| ComposeError::ShortBuf)
+    pub fn add_withdrawal(&mut self, withdrawal: &Nlri<Vec<u8>>)
+        -> Result<(), ComposeError>
+    {
+        //println!("add_withdrawal for {}", withdrawal.prefix().unwrap());
+        match *withdrawal {
+            Nlri::Unicast(b) => {
+                if b.is_v4() {
+                    if let Some(addpath_enabled) = self.addpath_enabled {
+                        if addpath_enabled != b.is_addpath() {
+                            return Err(ComposeError::IllegalCombination)
                         }
-                    ),+
-                        PathAttribute::Unimplemented(u) => {
-                            u.compose(target)
-                                .map_err(|_| ComposeError::ShortBuf)
-                        }
-                        PathAttribute::Invalid(_,_,_) => {
-                            Err(ComposeError::InvalidAttribute)
-                        }
+                    } else {
+                        self.addpath_enabled = Some(b.is_addpath());
                     }
-                }
 
-                pub fn typecode(&self) -> PathAttributeType {
-                    match self {
-                        $(
-                        PathAttribute::$name(_) =>
-                            PathAttributeType::$name
-                        ),+,
-                        PathAttribute::Unimplemented(u) => {
-                            PathAttributeType::Unimplemented(u.typecode())
-                        }
-                        PathAttribute::Invalid(_, tc, _) => {
-                            PathAttributeType::Invalid(*tc)
-                        }
+                    let new_bytes_num = withdrawal.compose_len();
+                    let new_total = self.total_pdu_len + new_bytes_num;
+                    if new_total > Self::MAX_PDU {
+                        return Err(ComposeError::PduTooLarge(new_total));
                     }
-                }
-            }
-
-            $(
-            impl From<$name> for PathAttribute {
-                fn from(pa: $name) -> PathAttribute {
-                    PathAttribute::$name(pa)
-                }
-            }
-            )+
-
-            impl From<UnimplementedPathAttribute> for PathAttribute {
-                fn from(u: UnimplementedPathAttribute) -> PathAttribute {
-                    PathAttribute::Unimplemented(u)
-                }
-            }
-
-            #[derive(Debug)]
-            pub enum WireformatPathAttribute<'a, Octs: Octets> {
-                $( $name(Parser<'a, Octs>, SessionConfig) ),+,
-                Unimplemented(UnimplementedWireformat<'a, Octs>),
-                Invalid(Flags, u8, Parser<'a, Octs>),
-            }
-
-
-            impl<'a, Octs: Octets> WireformatPathAttribute<'a, Octs> {
-                fn parse(parser: &mut Parser<'a, Octs>, sc: SessionConfig) 
-                    -> Result<WireformatPathAttribute<'a, Octs>, ParseError>
-                {
-                    let start_pos = parser.pos();
-                    let flags = parser.parse_u8()?;
-                    let typecode = parser.parse_u8()?;
-                    let (header_len, len) = match flags & 0x10 == 0x10 {
-                        true => (4, parser.parse_u16_be()? as usize),
-                        false => (3, parser.parse_u8()? as usize),
-                    };
-
-                    let mut pp = parser.parse_parser(len)?;
-
-                    let res = match typecode {
-                        $(
-                        $typecode => {
-                            if let Err(e) = $name::validate(
-                                flags.into(), &mut pp, sc
+                    self.withdrawals.push(withdrawal.clone());
+                    self.withdrawals_len += new_bytes_num;
+                    self.total_pdu_len = new_total;
+                } else {
+                    // Nlri::Unicast only holds IPv4 and IPv6, so this must be
+                    // IPv6.
+                    let new_bytes_num = match self.mp_unreach_nlri_builder {
+                        None => {
+                            let builder = MpUnreachNlriBuilder::new(
+                                AFI::Ipv6, SAFI::Unicast,
+                                b.is_addpath()
+                            );
+                            let res = builder.compose_len(withdrawal);
+                            self.mp_unreach_nlri_builder = Some(builder);
+                            res
+                            
+                        }
+                        Some(ref builder) => {
+                            if !builder.valid_combination(
+                                AFI::Ipv6, SAFI::Unicast, b.is_addpath()
                             ) {
-                                debug!("failed to parse path attribute: {e}");
-                                pp.seek(start_pos)?;
-                                WireformatPathAttribute::Invalid(
-                                    $flags.into(), $typecode, pp
-                                )
-                            } else {
-                                pp.seek(start_pos + header_len)?;
-                                WireformatPathAttribute::$name(pp, sc)
+                                // We are already constructing a
+                                // MP_UNREACH_NLRI but for a different
+                                // AFI,SAFI than the prefix in `withdrawal`,
+                                // or we are mixing addpath with non-addpath.
+                                return Err(ComposeError::IllegalCombination);
                             }
+                            builder.compose_len(withdrawal)
                         }
-                        ),+
-                        ,
-                        _ => WireformatPathAttribute::Unimplemented(
-                            UnimplementedWireformat::new(
-                                flags.into(), typecode, pp
-                            ))
                     };
 
-                    Ok(res)
-                }
+                    let new_total = self.total_pdu_len + new_bytes_num;
 
-                // XXX this method is the reason we have fn parse as part of
-                // the trait, forcing us the pass a SessionConfig to all of
-                // the parse() implementations.
-                pub fn to_owned(&self) -> Result<PathAttribute, ParseError> {
-                    match self {
-                        $(
-                        WireformatPathAttribute::$name(p, sc) => {
-                            Ok(PathAttribute::$name(
-                                $name::parse(&mut p.clone(), *sc)?
-                            ))
-                        }
-                        ),+,
-                        WireformatPathAttribute::Unimplemented(u) => {
-                            Ok(PathAttribute::Unimplemented(
-                                UnimplementedPathAttribute::new(
-                                    u.flags(),
-                                    u.typecode(),
-                                    u.value().to_vec()
-                                )
-                            ))
-                        },
-                        WireformatPathAttribute::Invalid(f, tc, p) => {
-                            Ok(PathAttribute::Invalid(
-                                *f, *tc, p.peek_all().to_vec()
-                            ))
-                        }
+                    if new_total > Self::MAX_PDU {
+                        return Err(ComposeError::PduTooLarge(new_total));
+                    }
+
+                    if let Some(ref mut builder) = self.mp_unreach_nlri_builder {
+                        builder.add_withdrawal(withdrawal);
+                        self.attributes_len += new_bytes_num;
+                        self.total_pdu_len = new_total;
+                    } else {
+                        // We always have Some builder at this point.
+                        unreachable!()
                     }
                 }
 
-                pub fn typecode(&self) -> PathAttributeType {
-                    match self {
-                        $(
-                            WireformatPathAttribute::$name(_,_) =>
-                                PathAttributeType::$name
-                        ),+,
-                        WireformatPathAttribute::Unimplemented(u) => {
-                            PathAttributeType::Unimplemented(u.typecode())
-                        }
-                        WireformatPathAttribute::Invalid(_, tc, _) => {
-                            PathAttributeType::Invalid(*tc)
-                        }
+            }
+            _ => todo!() // TODO use the MpUnreachNlriBuilder for these
+        };
+        Ok(())
+    }
+
+    pub fn withdrawals_from_iter<I>(&mut self, withdrawals: &mut Peekable<I>)
+        -> Result<(), ComposeError>
+    where I: Iterator<Item = Nlri<Vec<u8>>>
+    {
+        while let Some(w) = withdrawals.peek() {
+            match self.add_withdrawal(w) {
+                Ok(_) => { withdrawals.next(); }
+                Err(e) => return Err(e)
+            }
+        }
+        Ok(())
+    }
+
+    pub fn append_withdrawals(&mut self, withdrawals: &mut Vec<Nlri<Vec<u8>>>)
+        -> Result<(), ComposeError>
+    {
+        for (idx, w) in withdrawals.iter().enumerate() {
+            if let Err(e) = self.add_withdrawal(w) {
+
+                *withdrawals = withdrawals[idx..].to_vec();
+                return Err(e);
+            }
+        }
+        Ok(())
+    }
+
+    //--- Path Attributes
+
+    pub fn set_origin(&mut self, origin: OriginType)
+        -> Result<(), ComposeError>
+    {
+        if self.origin.is_none() {
+            // Check if this goes over the total PDU len. That will happen
+            // before we will go over the u16 for Total Path Attributes Length
+            // so we don't have to check for that.
+            let new_total = self.total_pdu_len + 4; // XXX should this 4 come
+                                                    // from a
+                                                    // `fn compose_len()` on
+                                                    // the attribute itself
+                                                    // perhaps?
+            if new_total > Self::MAX_PDU {
+                return Err(ComposeError::PduTooLarge(new_total));
+            } else {
+                self.total_pdu_len = new_total;
+                self.attributes_len += 4
+            }
+        }
+        self.origin = Some(new_pas::Origin::new(origin));
+        Ok(())
+    }
+
+    pub fn set_aspath<Octs>(&mut self , aspath: AsPath<Vec<u8>>)
+        -> Result<(), ComposeError>
+    {
+        if aspath.compose_len() > u16::MAX.into()  {
+            return Err(ComposeError::AttributeTooLarge(
+                 new_pas::PathAttributeType::AsPath, aspath.compose_len()
+            ));
+        }
+        if let Some(old_aspath) = &self.aspath {
+            let new_total = self.total_pdu_len
+                - old_aspath.compose_len()
+                + aspath.compose_len();
+            if new_total > Self::MAX_PDU {
+                return Err(ComposeError::PduTooLarge(new_total));
+            } else {
+                self.total_pdu_len -= old_aspath.compose_len();
+                self.attributes_len -= old_aspath.compose_len();
+            }
+        }
+        self.attributes_len += aspath.compose_len();
+        self.total_pdu_len += aspath.compose_len();
+        self.aspath = Some(aspath);
+        Ok(())
+    }
+
+    pub fn set_nexthop(&mut self, addr: IpAddr) -> Result<(), ComposeError> {
+        // Depending on the variant of `addr` we add/update either:
+        // - the conventional NEXT_HOP path attribute (IPv4); or
+        // - we update/create a MpReachNlriBuilder (IPv6)
+        match addr {
+            IpAddr::V4(a) => {
+                if self.nexthop.is_none() {
+                    // NEXT_HOP path attribute is 7 bytes long.
+                    let new_total = self.total_pdu_len + 7;
+                    if new_total > Self::MAX_PDU {
+                        return Err(ComposeError::PduTooLarge(new_total));
+                    } else {
+                        self.total_pdu_len = new_total;
+                        self.attributes_len += 7
                     }
+                    self.nexthop = Some(new_pas::NextHop::new(a));
                 }
             }
-
-            impl<'a, Octs: Octets> AsRef<[u8]> for WireformatPathAttribute<'a, Octs> {
-                fn as_ref(&self) -> &[u8] {
-                    match self {
-                        $(
-                            WireformatPathAttribute::$name(p, _) => {
-                                // this one is maybe fixed this way?
-                            p.peek_all()
-                        }
-                        ),+,
-                            WireformatPathAttribute::Unimplemented(u) => {
-                                u.value()
-                            }
-                            WireformatPathAttribute::Invalid(_, _, pp) => {
-                                pp.peek_all()
-                            }
-
-                    }
-                }
-            }
-
-            #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-            pub enum PathAttributeType {
-                $( $name ),+,
-                Unimplemented(u8),
-                Invalid(u8),
-            }
-            impl From<u8> for PathAttributeType {
-                fn from(code: u8) -> PathAttributeType {
-                    match code {
-                        $( $typecode => PathAttributeType::$name ),+,
-                        u => PathAttributeType::Unimplemented(u)
-                    }
-                }
-            }
-
-            impl From<PathAttributeType> for u8 {
-                fn from(pat: PathAttributeType) -> u8 {
-                    match pat {
-                        $( PathAttributeType::$name => $typecode ),+,
-                        PathAttributeType::Unimplemented(i) => i,
-                        PathAttributeType::Invalid(i) => i,
-                    }
-                }
-            }
-
-            $(
-            attribute!($name($data), $flags, $typecode);
-            )+
-        }
-    }
-
-    path_attributes!(
-        1   => Origin(OriginType), Flags::WELLKNOWN,
-        2   => AsPath(HopPath), Flags::WELLKNOWN,
-        3   => NextHop(Ipv4Addr), Flags::WELLKNOWN,
-        4   => MultiExitDisc(u32), Flags::OPT_NON_TRANS,
-        5   => LocalPref(u32), Flags::WELLKNOWN,
-        6   => AtomicAggregate(()), Flags::WELLKNOWN,
-        7   => Aggregator(AggregatorInfo), Flags::OPT_TRANS,
-        8   => Communities(StandardCommunitiesList), Flags::OPT_TRANS,
-        9   => OriginatorId(Ipv4Addr), Flags::OPT_NON_TRANS,
-        10  => ClusterList(ClusterIds), Flags::OPT_NON_TRANS,
-        14  => MpReachNlri(MpReachNlriBuilder), Flags::OPT_NON_TRANS,
-        15  => MpUnreachNlri(MpUnreachNlriBuilder), Flags::OPT_NON_TRANS,
-        16  => ExtendedCommunities(ExtendedCommunitiesList), Flags::OPT_TRANS,
-        17  => As4Path(HopPath), Flags::OPT_TRANS,
-        18  => As4Aggregator(AggregatorInfo), Flags::OPT_TRANS,
-        20  => Connector(Ipv4Addr), Flags::OPT_TRANS,
-        21  => AsPathLimit(AsPathLimitInfo), Flags::OPT_TRANS,
-        //22  => PmsiTunnel(todo), Flags::OPT_TRANS,
-        25  => Ipv6ExtendedCommunities(Ipv6ExtendedCommunitiesList), Flags::OPT_TRANS,
-        32  => LargeCommunities(LargeCommunitiesList), Flags::OPT_TRANS,
-        // 33 => BgpsecAsPath,
-        35 => Otc(Asn), Flags::OPT_TRANS,
-        //36 => BgpDomainPath(TODO), Flags:: , // https://datatracker.ietf.org/doc/draft-ietf-bess-evpn-ipvpn-interworking/06/
-        //40 => BgpPrefixSid(TODO), Flags::OPT_TRANS, // https://datatracker.ietf.org/doc/html/rfc8669#name-bgp-prefix-sid-attribute
-        128 => AttrSet(AttributeSet), Flags::OPT_TRANS,
-        255 => Reserved(ReservedRaw), Flags::OPT_TRANS,
-
-    );
-
-    #[derive(Debug, Eq, PartialEq)]
-    pub struct UnimplementedPathAttribute {
-        flags: Flags,
-        typecode: u8,
-        value: Vec<u8>,
-    }
-
-    impl UnimplementedPathAttribute {
-        pub fn new(flags: Flags, typecode: u8, value: Vec<u8>) -> Self {
-            Self { flags, typecode, value }
-        }
-
-        pub fn typecode(&self) -> u8 {
-            self.typecode
-        }
-
-        pub fn flags(&self) -> Flags {
-            self.flags
-        }
-
-        pub fn value(&self) -> &Vec<u8> {
-            &self.value
-        }
-    }
-
-    pub struct UnimplementedWireformat<'a, Octs: Octets> {
-        flags: Flags,
-        typecode: u8,
-        value: Parser<'a, Octs>,
-    }
-
-    impl<'a, Octs: Octets> fmt::Debug for UnimplementedWireformat<'a, Octs> {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, "{:08b} {} {:02x?}",
-               u8::from(self.flags()), self.typecode(), self.value()
-            )
-        }
-    }
-
-    impl<'a, Octs: Octets> UnimplementedWireformat<'a, Octs> {
-        pub fn new(flags: Flags, typecode: u8, value: Parser<'a, Octs>) -> Self {
-            Self { flags, typecode, value }
-        }
-        pub fn typecode(&self) -> u8 {
-            self.typecode
-        }
-
-        pub fn flags(&self) -> Flags {
-            self.flags
-        }
-
-        pub fn value(&self) -> &[u8] {
-            self.value.peek_all()
-        }
-
-        /* TODO move these to all (wireformat)pathattributes
-        /// Returns true if the optional flag is set.
-        pub fn is_optional(self) -> bool {
-            self.0 & 0x80 == 0x80
-        }
-
-        /// Returns true if the transitive bit is set.
-        pub fn is_transitive(&self) -> bool {
-            self.flags.is_transitive()
-        }
-
-        /// Returns true if the partial flag is set.
-        pub fn is_partial(&self) -> bool {
-            self.flags.is_partial()
-        }
-
-        /// Returns true if the extended length flag is set.
-        pub fn is_extended_length(&self) -> bool {
-            self.flags.is_extended_length()
-        }
-        */
-
-    }
-
-
-    pub trait Attribute: AttributeHeader {
-
-        fn compose_len(&self) -> usize {
-            self.header_len() + self.value_len()
-        }
-
-        fn is_extended(&self) -> bool {
-            self.value_len() > 255
-        }
-
-        fn header_len(&self) -> usize {
-            if self.is_extended() {
-                4
-            } else {
-                3
-            }
-        }
-
-        fn value_len(&self) -> usize;
-
-        fn compose<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            self.compose_header(target)?;
-            self.compose_value(target)
-        }
-
-        fn compose_header<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            if self.is_extended() {
-                target.append_slice(&[
-                    Self::FLAGS | Flags::EXTENDED_LEN,
-                    Self::TYPECODE,
-                ])?;
-                target.append_slice(
-                    &u16::try_from(self.value_len()).unwrap_or(u16::MAX)
-                    .to_be_bytes()
-                )
-            } else {
-                target.append_slice(&[
-                    Self::FLAGS,
-                    Self::TYPECODE,
-                    u8::try_from(self.value_len()).unwrap_or(u8::MAX)
-                ])
-            }
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>;
-
-        fn validate<Octs: Octets>(
-            flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            sc: SessionConfig
-        )
-            -> Result<(), ParseError>;
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, sc: SessionConfig) 
-            -> Result<Self, ParseError>
-        where Self: Sized;
-        
-    }
-
-    struct PathAttributes<'a, Octs> {
-        parser: Parser<'a, Octs>,
-        session_config: SessionConfig,
-    }
-
-    impl<'a, Octs> Clone for PathAttributes<'a, Octs> {
-        fn clone(&self) -> Self {
-            *self
-        }
-    }
-
-    impl<'a, Octs> Copy for PathAttributes<'a, Octs> { }
-
-    impl<'a, Octs: Octets> PathAttributes<'a, Octs> {
-        fn new(parser: Parser<'a, Octs>, session_config: SessionConfig)
-            -> PathAttributes<'_, Octs>
-        {
-            PathAttributes { parser, session_config }
-        }
-        
-        fn get(&self, pat: PathAttributeType)
-            -> Option<WireformatPathAttribute<'a, Octs>>
-        {
-            let mut iter = *self;
-            iter.find(|pa|
-                  //XXX We need Rust 1.70 for is_ok_and()
-                  //pa.as_ref().is_ok_and(|pa| pa.typecode() == pat)
-                  if let Ok(pa) = pa.as_ref() {
-                      pa.typecode() == pat
-                  } else {
-                      false
-                  }
-            ).map(|res| res.unwrap()) // res is Ok(pa), so we can unwrap.
-        }
-
-    }
-
-    impl<'a, Octs: Octets> Iterator for PathAttributes<'a, Octs> {
-        type Item = Result<WireformatPathAttribute<'a, Octs>, ParseError>;
-        fn next(&mut self) -> Option<Self::Item> {
-            if self.parser.remaining() == 0 {
-                return None;
-            }
-
-            let res = WireformatPathAttribute::parse(
-                &mut self.parser,
-                self.session_config
-            );
-            Some(res)
-        }
-    }
-
-    macro_rules! check_len_exact {
-        ($p:expr, $len:expr, $name:expr) => {
-            if $p.remaining() != $len {
-                Err(ParseError::form_error(
-                    "wrong length for $name, expected $len "
-                ))
-            } else {
-                Ok(())
-            }
-        }
-    }
-
-    //--- Origin
-
-    use crate::bgp::message::update::OriginType;
-
-    impl Attribute for Origin {
-        fn value_len(&self) -> usize { 1 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&[self.0.into()]) 
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<Origin, ParseError>
-        {
-            Ok(Origin(parser.parse_u8()?.into()))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            check_len_exact!(parser, 1, "ORIGIN")
-        }
-    }
-
-
-    //--- AsPath (see bgp::aspath)
-
-    impl Attribute for AsPath {
-        fn value_len(&self) -> usize {
-            self.0.to_as_path::<Vec<u8>>().unwrap().into_inner().len()
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-           target.append_slice(
-               self.0.to_as_path::<Vec<u8>>().unwrap().into_inner().as_ref()
-            )
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, sc: SessionConfig) 
-            -> Result<AsPath, ParseError>
-        {
-            // XXX reusing the old/existing AsPath here for the time being
-            let asp = crate::bgp::aspath::AsPath::new(
-                parser.peek_all().to_vec(),
-                sc.has_four_octet_asn()
-            ).map_err(|_| ParseError::form_error("invalid AS_PATH"))?;
-
-            Ok(AsPath(asp.to_hop_path()))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            let asn_size = if session_config.has_four_octet_asn() {
-                4
-            } else {
-                2
-            };
-            while parser.remaining() > 0 {
-                let segment_type = parser.parse_u8()?;
-                if !(1..=4).contains(&segment_type) {
-                    return Err(ParseError::form_error(
-                        "illegal segment type in AS_PATH"
-                    ));
-                }
-                let len = usize::from(parser.parse_u8()?); // ASNs in segment
-                parser.advance(len * asn_size)?; // ASNs.
-            }
-            Ok(())
-        }
-    }
-
-    //--- NextHop
-
-    impl Attribute for NextHop {
-        fn value_len(&self) -> usize { 4 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.octets())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<NextHop, ParseError>
-        {
-            Ok(NextHop(parse_ipv4addr(parser)?))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            check_len_exact!(parser, 4, "NEXT_HOP")
-        }
-    }
-
-    //--- MultiExitDisc
-
-    impl Attribute for MultiExitDisc {
-        fn value_len(&self) -> usize { 4 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.to_be_bytes()) 
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<MultiExitDisc, ParseError>
-        {
-            Ok(MultiExitDisc(parser.parse_u32_be()?))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            check_len_exact!(parser, 4, "MULTI_EXIT_DISC")
-        }
-    }
-
-    //--- LocalPref
-
-    impl Attribute for LocalPref {
-        fn value_len(&self) -> usize { 4 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.to_be_bytes()) 
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<LocalPref, ParseError>
-        {
-            Ok(LocalPref(parser.parse_u32_be()?))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            check_len_exact!(parser, 4, "LOCAL_PREF")
-        }
-    }
-
-    //--- AtomicAggregate
-
-    impl Attribute for AtomicAggregate {
-        fn value_len(&self) -> usize { 0 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, _target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            Ok(())
-        }
-
-        fn parse<Octs: Octets>(_parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<AtomicAggregate, ParseError>
-        {
-            Ok(AtomicAggregate(()))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            check_len_exact!(parser, 0, "ATOMIC_AGGREGATE")
-        }
-    }
-
-    //--- Aggregator
-
-    #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct AggregatorInfo {
-        asn: Asn,
-        address: Ipv4Addr,
-    }
-
-    impl AggregatorInfo {
-        fn new(asn: Asn, address: Ipv4Addr) -> AggregatorInfo {
-            AggregatorInfo { asn, address }
-        }
-        pub fn asn(&self) -> Asn {
-            self.asn
-        }
-        pub fn address(&self) -> Ipv4Addr {
-            self.address
-        }
-    }
-
-
-    impl Attribute for Aggregator {
-        // FIXME for legacy 2-byte ASNs, this should be 6.
-        // Should we pass a (&)SessionConfig to this method as well?
-        // Note that `fn compose_len` would then also need a SessionConfig,
-        // which, sort of makes sense anyway.
-        fn value_len(&self) -> usize { 
-            8
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.asn().to_raw())?;
-            target.append_slice(&self.0.address().octets())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, sc: SessionConfig) 
-            -> Result<Aggregator, ParseError>
-        {
-            let asn = if sc.has_four_octet_asn() {
-                Asn::from_u32(parser.parse_u32_be()?)
-            } else {
-                Asn::from_u32(parser.parse_u16_be()?.into())
-            };
-
-            let address = parse_ipv4addr(parser)?;
-            Ok(Aggregator(AggregatorInfo::new(asn, address)))
-        }
-
-        fn validate<Octs: Octets>(
-            flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            if flags != Self::FLAGS.into() {
-                    return Err(ParseError::form_error("invalid flags"));
-            }
-            if session_config.has_four_octet_asn() {
-                check_len_exact!(parser, 8, "AGGREGATOR")?;
-            } else {
-                check_len_exact!(parser, 6, "AGGREGATOR")?;
-            }
-            Ok(())
-        }
-    }
-
-    //--- Communities
-    
-    use crate::bgp::communities::StandardCommunity;
-    #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct StandardCommunitiesList {
-        communities: Vec<StandardCommunity>
-    }
-
-    impl StandardCommunitiesList {
-        fn new(communities: Vec<StandardCommunity>)
-            -> StandardCommunitiesList
-        {
-            StandardCommunitiesList {communities }
-        }
-
-        pub fn communities(&self) -> &Vec<StandardCommunity> {
-            &self.communities
-        }
-    }
-
-
-    impl Attribute for Communities {
-        fn value_len(&self) -> usize { 
-            self.0.communities.len() * 4
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            for c in &self.0.communities {
-                target.append_slice(&c.to_raw())?;
-            }
-            Ok(())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<Communities, ParseError>
-        {
-            let mut communities = Vec::with_capacity(parser.remaining() / 4);
-            while parser.remaining() > 0 {
-                communities.push(parser.parse_u32_be()?.into());
-            }
-            Ok(Communities(StandardCommunitiesList::new(communities)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            if parser.remaining() % 4 != 0 {
-                return Err(ParseError::form_error(
-                    "unexpected length for COMMUNITIES"
-                ));
-            }
-            Ok(())
-        }
-    }
-
-    //--- OriginatorId
-
-    impl Attribute for OriginatorId {
-        fn value_len(&self) -> usize { 4 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.octets())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<OriginatorId, ParseError>
-        {
-            Ok(OriginatorId(parse_ipv4addr(parser)?))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            check_len_exact!(parser, 4, "ORIGINATOR_ID")
-        }
-    }
-
-    //--- ClusterList
-    
-    #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct BgpIdentifier([u8; 4]);
-
-    impl From<[u8; 4]> for BgpIdentifier {
-        fn from(raw: [u8; 4]) -> BgpIdentifier {
-            BgpIdentifier(raw)
-        }
-    }
-
-    /*
-    impl From<u32> for BgpIdentifier {
-        fn from(raw: u32) -> BgpIdentifier {
-            BgpIdentifier(raw.to_be_bytes())
-        }
-    }
-    */
-
-    #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct ClusterIds {
-        cluster_ids: Vec<BgpIdentifier>
-    }
-
-    impl ClusterIds {
-        fn new(cluster_ids: Vec<BgpIdentifier>) -> ClusterIds {
-            ClusterIds {cluster_ids }
-        }
-        pub fn cluster_ids(&self) -> &Vec<BgpIdentifier> {
-            &self.cluster_ids
-        }
-    }
-
-
-    impl Attribute for ClusterList {
-        fn value_len(&self) -> usize { 
-            self.0.cluster_ids.len() * 4
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            for c in &self.0.cluster_ids {
-                target.append_slice(&c.0)?;
-            }
-            Ok(())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<ClusterList, ParseError>
-        {
-            let mut cluster_ids = Vec::with_capacity(parser.remaining() / 4);
-            while parser.remaining() > 0 {
-                cluster_ids.push(parser.parse_u32_be()?.to_be_bytes().into());
-            }
-            Ok(ClusterList(ClusterIds::new(cluster_ids)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            if parser.remaining() % 4 != 0 {
-                return Err(ParseError::form_error(
-                    "unexpected length for CLUSTER_LIST"
-                ));
-            }
-            Ok(())
-        }
-    }
-
-    //--- MpReachNlri
-    impl Attribute for MpReachNlri {
-        fn value_len(&self) -> usize { 
-            self.0.value_len()
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            self.0.compose_value(target)
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, sc: SessionConfig) 
-            -> Result<MpReachNlri, ParseError>
-        {
-            let afi: AFI = parser.parse_u16_be()?.into();
-            let safi: SAFI = parser.parse_u8()?.into();
-            let nexthop = crate::bgp::types::NextHop::parse(parser, afi, safi)?;
-            parser.advance(1)?; // reserved byte
-            let mut builder = MpReachNlriBuilder::new(
-                afi, safi, nexthop, sc.addpath_enabled()
-            );
-            let nlri_iter = crate::bgp::message::update::Nlris::new(
-                *parser,
-                sc,
-                afi,
-                safi
-            ).iter();
-
-            // FIXME
-            // MpReachNlriBuilder works with Nlri<Vec<u8>>
-            // but that is somewhat limiting. Here, we reuse the existing
-            // Nlris iter from bgp::message::update, where the iterator
-            // returns Item = Nlri<Octs::Range<'_>>.
-            // Perhaps add_announcement should take Nlri<T> where T:
-            // OctetsInto<Vec<u8>> or something like that?
-            use crate::bgp::message::nlri::{Nlri, BasicNlri};
-            for nlri in nlri_iter {
-                match nlri {
-                    Nlri::Unicast(b) => {
-                        builder.add_announcement(
-                            &Nlri::Unicast(BasicNlri {
-                                prefix: b.prefix,
-                                path_id: b.path_id
-                            })
-                        )
-                    },
-                    _ => unimplemented!()
-                }
-            }
-            Ok(MpReachNlri(builder))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            // We only check for the bare minimum here, as most checks are
-            // better done upon (creation of the) Nlri iterator based on the
-            // value of this path attribute.
-            if parser.remaining() < 5 {
-                return Err(ParseError::form_error(
-                    "length for MP_REACH_NLRI less than minimum"
-                ))
-            }
-            Ok(())
-        }
-    }
-
-    //--- MpUnreachNlri
-    impl Attribute for MpUnreachNlri {
-        fn value_len(&self) -> usize { 
-            self.0.value_len()
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            self.0.compose_value(target)
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, sc: SessionConfig) 
-            -> Result<MpUnreachNlri, ParseError>
-        {
-            let afi: AFI = parser.parse_u16_be()?.into();
-            let safi: SAFI = parser.parse_u8()?.into();
-
-            let mut builder = MpUnreachNlriBuilder::new(
-                afi, safi, sc.addpath_enabled()
-            );
-            let nlri_iter = crate::bgp::message::update::Nlris::new(
-                *parser,
-                sc,
-                afi,
-                safi
-            ).iter();
-
-            // FIXME
-            // see note at MpReachNlri above
-            use crate::bgp::message::nlri::{Nlri, BasicNlri};
-            for nlri in nlri_iter {
-                match nlri {
-                    Nlri::Unicast(b) => {
-                        builder.add_withdrawal(
-                            &Nlri::Unicast(BasicNlri {
-                                prefix: b.prefix,
-                                path_id: b.path_id
-                            })
+            IpAddr::V6(a) => {
+                if let Some(ref mut builder) = self.mp_reach_nlri_builder {
+                    // Given that we have a builder, there is already either a
+                    // NextHop::Ipv6 or a NextHop::Ipv6LL. Setting the
+                    // non-link-local address does not change the length of
+                    // the next hop part, so there is no need to update the
+                    // total_pdu_len and the attributes_len.
+                    builder.set_nexthop(a);
+                } else {
+                    let builder = MpReachNlriBuilder::new(
+                        AFI::Ipv6, SAFI::Unicast, NextHop::Ipv6(a), false
                         );
-                    },
-                    _ => unimplemented!()
+                    let new_bytes = builder.compose_len_empty();
+                    let new_total = self.total_pdu_len + new_bytes;
+                    if new_total > Self::MAX_PDU {
+                        return Err(ComposeError::PduTooLarge(new_total));
+                    }
+                    self.attributes_len += new_bytes;
+                    self.total_pdu_len = new_total;
+                    self.mp_reach_nlri_builder = Some(builder);
                 }
             }
-
-            Ok(MpUnreachNlri(builder))
         }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            // We only check for the bare minimum here, as most checks are
-            // better done upon (creation of the) Nlri iterator based on the
-            // value of this path attribute.
-            if parser.remaining() < 3 {
-                return Err(ParseError::form_error(
-                    "length for MP_UNREACH_NLRI less than minimum"
-                ))
-            }
-            Ok(())
-        }
+        Ok(())
     }
 
-    //--- ExtendedCommunities
-    
-    use crate::bgp::communities::ExtendedCommunity;
-    #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct ExtendedCommunitiesList {
-        communities: Vec<ExtendedCommunity>
-    }
-
-    impl ExtendedCommunitiesList {
-        fn new(communities: Vec<ExtendedCommunity>)
-            -> ExtendedCommunitiesList
-        {
-            ExtendedCommunitiesList {communities }
-        }
-
-        pub fn communities(&self) -> &Vec<ExtendedCommunity> {
-            &self.communities
-        }
-    }
-
-
-    impl Attribute for ExtendedCommunities {
-        fn value_len(&self) -> usize { 
-            self.0.communities.len() * 8
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            for c in &self.0.communities {
-                target.append_slice(&c.to_raw())?;
-            }
-            Ok(())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<ExtendedCommunities, ParseError>
-        {
-            let mut communities = Vec::with_capacity(parser.remaining() / 8);
-            let mut buf = [0u8; 8];
-            while parser.remaining() > 0 {
-                parser.parse_buf(&mut buf)?;
-                communities.push(buf.into());
-            }
-            Ok(ExtendedCommunities(ExtendedCommunitiesList::new(communities)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            if parser.remaining() % 8 != 0 {
-                return Err(ParseError::form_error(
-                    "unexpected length for EXTENDED_COMMUNITIES"
-                ));
-            }
-            Ok(())
-        }
-    }
-
-    //--- As4Path (see bgp::aspath)
-
-    impl Attribute for As4Path {
-        fn value_len(&self) -> usize {
-            self.0.to_as_path::<Vec<u8>>().unwrap().into_inner().len()
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(
-                self.0.to_as_path::<Vec<u8>>().unwrap().into_inner().as_ref()
-            )
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, sc: SessionConfig) 
-            -> Result<As4Path, ParseError>
-        {
-            // XXX Same as with AsPath, reusing the old/existing As4Path here
-            let asp = crate::bgp::aspath::AsPath::new(
-                parser.peek_all().to_vec(),
-                sc.has_four_octet_asn()
-            ).map_err(|_| ParseError::form_error("invalid AS4_PATH"))?;
-            Ok(As4Path(asp.to_hop_path()))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            while parser.remaining() > 0 {
-                let segment_type = parser.parse_u8()?;
-                if !(1..=4).contains(&segment_type) {
-                    return Err(ParseError::form_error(
-                        "illegal segment type in AS4_PATH"
-                    ));
-                }
-                let len = usize::from(parser.parse_u8()?); // ASNs in segment
-                parser.advance(len * 4)?; // ASNs.
-            }
-            Ok(())
-        }
-
-    }
-
-    //--- As4Aggregator 
- 
-    impl Attribute for As4Aggregator {
-        fn value_len(&self) -> usize { 8 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.asn().to_raw())?;
-            target.append_slice(&self.0.address().octets())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<As4Aggregator, ParseError>
-        {
-            let asn = Asn::from_u32(parser.parse_u32_be()?);
-            let address = parse_ipv4addr(parser)?;
-            Ok(As4Aggregator(AggregatorInfo::new(asn, address)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            check_len_exact!(parser, 8, "AS4_AGGREGATOR")
-        }
-    }
-
-
-    //--- Connector (deprecated)
-
-    impl Attribute for Connector {
-        fn value_len(&self) -> usize { 4 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.octets())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<Connector, ParseError>
-        {
-            Ok(Connector(parse_ipv4addr(parser)?))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _sc: SessionConfig
-        )
-            -> Result<(), ParseError>
-        {
-            check_len_exact!(parser, 4, "CONNECTOR")
-        }
-    }
-
-    //--- AsPathLimit (deprecated)
-
-    #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct AsPathLimitInfo {
-        upper_bound: u8,
-        attacher: Asn,
-    }
-
-    impl AsPathLimitInfo {
-        pub fn new(upper_bound: u8, attacher: Asn) -> AsPathLimitInfo {
-            AsPathLimitInfo { upper_bound, attacher }
-        }
-    }
-
-    impl Attribute for AsPathLimit {
-        fn value_len(&self) -> usize { 5 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&[self.0.upper_bound])?;
-            target.append_slice(&self.0.attacher.to_raw())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<AsPathLimit, ParseError>
-        {
-            let info = AsPathLimitInfo {
-                upper_bound: parser.parse_u8()?,
-                attacher: Asn::from_u32(parser.parse_u32_be()?)
+    pub fn set_nexthop_ll(&mut self, addr: Ipv6Addr)
+        -> Result<(), ComposeError>
+    {
+        // XXX We could/should check for addr.is_unicast_link_local() once
+        // that lands in stable.
+        
+        if let Some(ref mut builder) = self.mp_reach_nlri_builder {
+            let new_bytes = match builder.get_nexthop() {
+                NextHop::Ipv6(_) => 16,
+                NextHop::Ipv6LL(_,_) => 0,
+                _ => unreachable!()
             };
 
-            Ok(AsPathLimit(info))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _sc: SessionConfig
-        )
-        -> Result<(), ParseError>
-        {
-            check_len_exact!(parser, 5, "AS_PATHLIMIT")
-        }
-    }
-
-    //--- Ipv6ExtendedCommunities
-    
-    use crate::bgp::communities::Ipv6ExtendedCommunity;
-    #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct Ipv6ExtendedCommunitiesList {
-        communities: Vec<Ipv6ExtendedCommunity>
-    }
-
-    impl Ipv6ExtendedCommunitiesList {
-        fn new(communities: Vec<Ipv6ExtendedCommunity>)
-            -> Ipv6ExtendedCommunitiesList
-        {
-            Ipv6ExtendedCommunitiesList {communities }
-        }
-
-        pub fn communities(&self) -> &Vec<Ipv6ExtendedCommunity> {
-            &self.communities
-        }
-    }
-
-
-    impl Attribute for Ipv6ExtendedCommunities {
-        fn value_len(&self) -> usize { 
-            self.0.communities.len() * 20
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            for c in &self.0.communities {
-                target.append_slice(&c.to_raw())?;
+            let new_total = self.total_pdu_len + new_bytes;
+            if new_total > Self::MAX_PDU {
+                return Err(ComposeError::PduTooLarge(new_total));
             }
-            Ok(())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<Ipv6ExtendedCommunities, ParseError>
-        {
-            let mut communities = Vec::with_capacity(parser.remaining() / 20);
-            let mut buf = [0u8; 20];
-            while parser.remaining() > 0 {
-                parser.parse_buf(&mut buf)?;
-                communities.push(buf.into());
+            builder.set_nexthop_ll(addr);
+            self.attributes_len += new_bytes;
+            self.total_pdu_len = new_total;
+        } else {
+            let builder = MpReachNlriBuilder::new(
+                AFI::Ipv6, SAFI::Unicast, NextHop::Ipv6LL(0.into(), addr),
+                false
+            );
+            let new_bytes = builder.compose_len_empty();
+            let new_total = self.total_pdu_len + new_bytes;
+            if new_total > Self::MAX_PDU {
+                return Err(ComposeError::PduTooLarge(new_total));
             }
-            Ok(Ipv6ExtendedCommunities(Ipv6ExtendedCommunitiesList::new(communities)))
+            self.attributes_len += new_bytes;
+            self.total_pdu_len = new_total;
+            self.mp_reach_nlri_builder = Some(builder);
         }
 
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            if parser.remaining() % 20 != 0 {
-                return Err(ParseError::form_error(
-                    "unexpected length for IPV6_EXTENDED_COMMUNITIES"
-                ));
-            }
-            Ok(())
-        }
+        Ok(())
     }
 
-    //--- LargeCommunities
-    
-    use crate::bgp::communities::LargeCommunity;
-    #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct LargeCommunitiesList {
-        communities: Vec<LargeCommunity>
+    pub fn set_multi_exit_disc(&mut self, med: new_pas::MultiExitDisc)
+    -> Result<(), ComposeError>
+    {
+        let new_bytes = med.compose_len();
+        let new_total = self.total_pdu_len + new_bytes;
+        if new_total > Self::MAX_PDU {
+            return Err(ComposeError::PduTooLarge(new_total));
+        }
+        self.multi_exit_disc = Some(med);
+        self.total_pdu_len = new_total;
+        self.attributes_len += new_bytes;
+        Ok(())
     }
 
-    impl LargeCommunitiesList {
-        fn new(communities: Vec<LargeCommunity>)
-            -> LargeCommunitiesList
-        {
-            LargeCommunitiesList {communities }
+    pub fn set_local_pref(&mut self, local_pref: new_pas::LocalPref)
+    -> Result<(), ComposeError>
+    {
+        let new_bytes = local_pref.compose_len();
+        let new_total = self.total_pdu_len + new_bytes;
+        if new_total > Self::MAX_PDU {
+            return Err(ComposeError::PduTooLarge(new_total));
         }
-
-        pub fn communities(&self) -> &Vec<LargeCommunity> {
-            &self.communities
-        }
-    }
-
-
-    impl Attribute for LargeCommunities {
-        fn value_len(&self) -> usize { 
-            self.0.communities.len() * 12
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            for c in &self.0.communities {
-                target.append_slice(&c.to_raw())?;
-            }
-            Ok(())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<LargeCommunities, ParseError>
-        {
-            let mut communities = Vec::with_capacity(parser.remaining() / 12);
-            let mut buf = [0u8; 12];
-            while parser.remaining() > 0 {
-                parser.parse_buf(&mut buf)?;
-                communities.push(buf.into());
-            }
-            Ok(LargeCommunities(LargeCommunitiesList::new(communities)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _session_config: SessionConfig
-        ) -> Result<(), ParseError> {
-            if parser.remaining() % 12 != 0 {
-                return Err(ParseError::form_error(
-                    "unexpected length for LARGE_COMMUNITIES"
-                ));
-            }
-            Ok(())
-        }
-    }
-
-    //--- Otc 
- 
-    impl Attribute for Otc {
-        fn value_len(&self) -> usize { 4 }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.to_raw())
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<Otc, ParseError>
-        {
-            Ok(Otc(Asn::from_u32(parser.parse_u32_be()?)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _sc: SessionConfig
-        )
-        -> Result<(), ParseError>
-        {
-            check_len_exact!(parser, 4, "OTC")
-        }
-    }
-
-    //--- AttributeSet
-
-    #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct AttributeSet {
-        origin: Asn,
-        attributes: Vec<u8>,
-    }
-
-    impl AttributeSet {
-        pub fn new(origin: Asn, attributes: Vec<u8>) -> AttributeSet {
-            AttributeSet { origin, attributes }
-        }
-    }
- 
-    impl Attribute for AttrSet {
-        fn value_len(&self) -> usize {
-            4 + self.0.attributes.len()
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.origin.to_raw())?;
-            target.append_slice(&self.0.attributes)
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<AttrSet, ParseError>
-        {
-            let origin = Asn::from_u32(parser.parse_u32_be()?);
-            let attributes = parser.peek_all().to_vec();
-            Ok(AttrSet(AttributeSet::new(origin, attributes)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            parser: &mut Parser<'_, Octs>,
-            _sc: SessionConfig
-        )
-        -> Result<(), ParseError>
-        {
-            // XXX we do not validate the actual content (i.e. the attributes)
-            // here. Whoever wishes to process these will need to iterate over
-            // them with a PathAttributes anyway.
-            if parser.remaining() < 4 {
-                return Err(ParseError::form_error(
-                    "length for ATTR_SET less than minimum"
-                ))
-            }
-            Ok(())
-        }
-    }
-
-    //--- ReservedRaw
-
-    #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct ReservedRaw {
-        raw: Vec<u8>,
-    }
-
-    impl ReservedRaw {
-        pub fn new(raw: Vec<u8>) -> ReservedRaw {
-            ReservedRaw { raw }
-        }
-    }
- 
-    impl Attribute for Reserved {
-        fn value_len(&self) -> usize {
-            self.0.raw.len()
-        }
-
-        fn compose_value<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            target.append_slice(&self.0.raw)
-        }
-
-        fn parse<Octs: Octets>(parser: &mut Parser<Octs>, _sc: SessionConfig) 
-            -> Result<Reserved, ParseError>
-        {
-            let raw = parser.peek_all().to_vec();
-            Ok(Reserved(ReservedRaw::new(raw)))
-        }
-
-        fn validate<Octs: Octets>(
-            _flags: Flags,
-            _parser: &mut Parser<'_, Octs>,
-            _sc: SessionConfig
-        )
-        -> Result<(), ParseError>
-        {
-            // Not anything we can validate here, really.
-            Ok(())
-        }
+        self.local_pref = Some(local_pref);
+        self.total_pdu_len = new_total;
+        self.attributes_len += new_bytes;
+        Ok(())
     }
 
 
-    //--- Unimplemented
-    // 
-    // XXX implementing the Attribute trait requires to implement the
-    // AttributeHeader trait to be implemented as well. But, we have no const
-    // FLAGS and TYPECODE for an UnimplementedPathAttribute, so that does not
-    // fly.
-    //
-    // Let's try to go without the trait.
+    //--- Announcements
 
-    impl UnimplementedPathAttribute {
-        fn compose<Target: OctetsBuilder>(&self, target: &mut Target)
-            -> Result<(), Target::AppendError>
-        {
-            let len = self.value().len();
-            target.append_slice(
-                &[self.flags().into(), self.typecode()]
-            )?;
-            if self.flags().is_extended_length() {
-                target.append_slice(
-                    &u16::try_from(len).unwrap_or(u16::MAX)
-                    .to_be_bytes()
-                )?;
-            } else {
-                target.append_slice(&[
-                    u8::try_from(len).unwrap_or(u8::MAX)
-                ])?;
-            }
-            target.append_slice(self.value())
-        }
-    }
+    pub fn add_announcement(&mut self, announcement: &Nlri<Vec<u8>>)
+        -> Result<(), ComposeError>
+    {
+        match *announcement {
+            Nlri::Unicast(b) => {
+                if b.is_v4() {
+                    if let Some(addpath_enabled) = self.addpath_enabled {
+                        if addpath_enabled != b.is_addpath() {
+                            return Err(ComposeError::IllegalCombination)
+                        }
+                    } else {
+                        self.addpath_enabled = Some(b.is_addpath());
+                    }
 
+                    let new_bytes_num = announcement.compose_len();
+                    let new_total = self.total_pdu_len + new_bytes_num;
+                    if new_total > Self::MAX_PDU {
+                        return Err(ComposeError::PduTooLarge(new_total));
+                    }
+                    self.announcements.push(announcement.clone());
+                    self.announcements_len += new_bytes_num;
+                    self.total_pdu_len = new_total;
+                    
+                } else {
+                    // Nlri::Unicast only holds IPv4 and IPv6, so this must be
+                    // IPv6 unicast.
 
-//------------ Tests ---------------------------------------------------------
+                    let new_bytes_num = match self.mp_reach_nlri_builder {
+                        None => {
+                            let nexthop = NextHop::Ipv6(0.into());
+                            let builder = MpReachNlriBuilder::new(
+                                AFI::Ipv6, SAFI::Unicast,
+                                nexthop,
+                                b.is_addpath()
+                            );
+                            let res = builder.compose_len_empty() + 
+                                builder.compose_len(announcement);
+                            self.mp_reach_nlri_builder = Some(builder);
+                            res
+                            
+                        }
+                        Some(ref builder) => {
+                            //TODO we should allow multicast here, but then we
+                            //should also check whether a prefix is_multicast.
+                            //Similarly, should we check for is_unicast?
+                            //Or should we handle anything other than unicast
+                            //in the outer match anyway, as that is never a
+                            //conventional announcement anyway?
+                            
+                            if !builder.valid_combination(
+                                AFI::Ipv6, SAFI::Unicast, b.is_addpath()
+                            ) {
+                                return Err(ComposeError::IllegalCombination);
+                            }
 
-#[cfg(test)]
-    mod tests {
-        use super::*;
-        use crate::asn::Asn;
-        use crate::bgp::communities::Wellknown;
-        use crate::bgp::message::nlri::Nlri;
-        use crate::bgp::message::update::NextHop;
+                            builder.compose_len(announcement)
+                        }
+                    };
 
-        // FIXME switch to use ::new() for all attributes instead of relying
-        // on non public newtype constructor
-        #[test]
-        fn wireformat_to_owned_and_back() {
-            use super::PathAttribute as PA;
-            fn check(raw: Vec<u8>, owned: PathAttribute) {
-                let mut parser = Parser::from_ref(&raw);
-                let sc = SessionConfig::modern();
-                let pa = WireformatPathAttribute::parse(&mut parser, sc)
-                    .unwrap();
-                assert_eq!(owned, pa.to_owned().unwrap());
-                let mut target = Vec::new();
-                owned.compose(&mut target).unwrap();
-                assert_eq!(target, raw);
-            }
+                    let new_total = self.total_pdu_len + new_bytes_num;
 
-            check(vec![0x40, 0x01, 0x01, 0x00], PA::Origin(Origin(0.into())));
+                    if new_total > Self::MAX_PDU {
+                        return Err(ComposeError::PduTooLarge(new_total));
+                    }
 
-            check(
-                vec![0x40, 0x02, 10,
-                0x02, 0x02, // SEQUENCE of length 2
-                0x00, 0x00, 0x00, 100,
-                0x00, 0x00, 0x00, 200,
-                ],
-                PA::AsPath(AsPath(HopPath::from(vec![
-                    Asn::from_u32(100),
-                    Asn::from_u32(200)]
-                )))
-            );
-
-            check(
-                vec![0x40, 0x03, 0x04, 0x01, 0x02, 0x03, 0x04],
-                PA::NextHop(NextHop("1.2.3.4".parse().unwrap()))
-            );
-
-            check(
-                vec![0x80, 0x04, 0x04, 0x00, 0x00, 0x00, 0xff],
-                PA::MultiExitDisc(MultiExitDisc::new(255))
-            );
-
-            check(
-                vec![0x40, 0x05, 0x04, 0x00, 0x00, 0x00, 0x0a],
-                PA::LocalPref(LocalPref::new(10))
-            );
-
-            check(
-                vec![0x40, 0x06, 0x00],
-                PA::AtomicAggregate(AtomicAggregate(()))
-            );
-            
-            check(
-                vec![
-                    0xc0, 0x07, 0x08, 0x00, 0x00, 0x00, 0x65, 0xc6,
-                    0x33, 0x64, 0x01
-                ],
-                PA::Aggregator(Aggregator(AggregatorInfo::new(
-                        Asn::from_u32(101),
-                        "198.51.100.1".parse().unwrap()
-                )))
-            );
-
-            check(
-                vec![
-                    0xc0, 0x08, 0x10, 0x00, 0x2a, 0x02, 0x06, 0xff,
-                    0xff, 0xff, 0x01, 0xff, 0xff, 0xff, 0x02, 0xff,
-                    0xff, 0xff, 0x03
-                ],
-                PA::Communities(Communities(StandardCommunitiesList::new(
-                    vec!["AS42:518".parse().unwrap(),
-                        Wellknown::NoExport.into(),
-                        Wellknown::NoAdvertise.into(),
-                        Wellknown::NoExportSubconfed.into(),
-                    ]
-                )))
-            );
-
-            check(
-                vec![0x80, 0x09, 0x04, 0x0a, 0x00, 0x00, 0x04],
-                OriginatorId("10.0.0.4".parse().unwrap()).into()
-            );
-
-            check(
-                vec![0x80, 0x0a, 0x04, 0x0a, 0x00, 0x00, 0x03],
-                ClusterList(ClusterIds::new(
-                        vec![[10, 0, 0, 3].into()]
-                )).into()
-            );
-            
-            check(
-                vec![
-                    0x80, 0x0e, 0x1c,
-                    0x00, 0x02, 0x01,
-                    0x10,
-                    0x20, 0x01, 0x0d, 0xb8, 0x00, 0x00, 0x00, 0x00,
-                    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x34,
-                    0x00,
-                    0x30, 0x20, 0x01, 0x0d, 0xb8, 0xaa, 0xbb
-                ],
-                {
-                let mut builder = MpReachNlriBuilder::new(
-                    AFI::Ipv6,
-                    SAFI::Unicast,
-                    NextHop::Ipv6("2001:db8::1234".parse().unwrap()),
-                    false // no addpath
-                );
-                builder.add_announcement(
-                    &Nlri::unicast_from_str("2001:db8:aabb::/48").unwrap()
-                );
-
-                MpReachNlri(builder).into()
+                    if let Some(ref mut builder) = self.mp_reach_nlri_builder {
+                        builder.add_announcement(announcement);
+                        self.attributes_len += new_bytes_num;
+                        self.total_pdu_len = new_total;
+                    } else {
+                        // We always have Some builder at this point.
+                        unreachable!()
+                    }
                 }
-            );
+            }
+            _ => todo!() // TODO use MpReachNlriBuilder once we have that.
+        }
 
-            check(
-                vec![
-                    0x80, 0x0f, 0x27, 0x00, 0x02, 0x01, 0x40, 0x20,
-                    0x01, 0x0d, 0xb8, 0xff, 0xff, 0x00, 0x00, 0x40,
-                    0x20, 0x01, 0x0d, 0xb8, 0xff, 0xff, 0x00, 0x01,
-                    0x40, 0x20, 0x01, 0x0d, 0xb8, 0xff, 0xff, 0x00,
-                    0x02, 0x40, 0x20, 0x01, 0x0d, 0xb8, 0xff, 0xff,
-                    0x00, 0x03
-                ],
-                {
-                let mut builder = MpUnreachNlriBuilder::new(
-                    AFI::Ipv6,
-                    SAFI::Unicast,
-                    false // no addpath
-                );
-                [
-                    "2001:db8:ffff::/64",
-                    "2001:db8:ffff:1::/64",
-                    "2001:db8:ffff:2::/64",
-                    "2001:db8:ffff:3::/64",
-                ].into_iter().for_each(|s|{
-                    builder.add_withdrawal(
-                        &Nlri::unicast_from_str(s).unwrap()
+        Ok(())
+    }
+
+    pub fn announcements_from_iter<I>(
+        &mut self, announcements: &mut Peekable<I>
+    ) -> Result<(), ComposeError>
+        where I: Iterator<Item = Nlri<Vec<u8>>>
+    {
+        while let Some(a) = announcements.peek() {
+            match self.add_announcement(a) {
+                Ok(_) => { announcements.next(); }
+                Err(e) => return Err(e)
+            }
+        }
+        Ok(())
+    }
+
+    //--- Standard communities
+    
+    pub fn add_community(&mut self, community: StandardCommunity)
+        -> Result<(), ComposeError>
+    {
+
+        if let Some(ref mut builder) = self.standard_communities_builder {
+            let new_bytes = builder.compose_len(community);
+            let new_total = self.total_pdu_len + new_bytes;
+            if new_total > Self::MAX_PDU {
+                return Err(ComposeError::PduTooLarge(new_total));
+            }
+            builder.add_community(community);
+            self.attributes_len += new_bytes;
+            self.total_pdu_len = new_total;
+        } else {
+            let mut builder = StandardCommunitiesBuilder::new();
+            let new_bytes = builder.compose_len_empty()
+                + builder.compose_len(community); 
+            let new_total = self.total_pdu_len + new_bytes;
+            if new_total > Self::MAX_PDU {
+                return Err(ComposeError::PduTooLarge(new_total));
+            }
+            builder.add_community(community);
+            self.standard_communities_builder = Some(builder);
+            self.attributes_len += new_bytes;
+            self.total_pdu_len = new_total;
+        }
+
+        Ok(())
+    }
+
+}
+
+impl<Target: OctetsBuilder + AsMut<[u8]>> UpdateBuilder<Target>
+where Infallible: From<<Target as OctetsBuilder>::AppendError>
+{
+    #[deprecated]
+    pub fn build_acs(mut self, acs: AttrChangeSet)
+        -> Result<Target, ComposeError>
+    {
+        // Withdrawals
+        let withdraw_len = 0_usize;
+        // placeholder
+        let _ = self.target.append_slice(&(withdraw_len as u16).to_be_bytes());
+        //self.target.as_mut()[19..=20].copy_from_slice(
+        //    &(withdraw_len as u16).to_be_bytes()
+        //);
+        // TODO actual withdrawals
+
+
+        // Path Attributes
+            // flags (from msb to lsb):
+            // optional
+            // transitive
+            // partial
+            // extended_length (2 octet length)
+        
+        let mut total_pa_len = 0_usize;
+        // Total Path Attribute len place holder:
+        let _ = self.target.append_slice(&[0x00, 0x00]);
+
+        if let Some(origin) = acs.origin_type.into_opt() {
+            let attr_flags = 0b0100_0000;
+            let attr_typecode = PathAttributeType::Origin.into();
+            let attr_len = 1_u8; 
+            let _ = self.target.append_slice(
+                &[attr_flags, attr_typecode, attr_len, origin.into()]);
+            total_pa_len += 2 + 1 + usize::from(attr_len);
+        }
+
+        if let Some(as_path) = acs.as_path.into_opt() {
+            let attr_flags = 0b0101_0000;
+            let attr_typecode = PathAttributeType::AsPath.into();
+            let asp = as_path.into_inner();
+            let attr_len = asp.len();
+            if u16::try_from(attr_len).is_err() {
+                return Err(ComposeError::AttributeTooLarge(
+                    PathAttributeType::AsPath,
+                    attr_len
+                ));
+            }
+            let _ = self.target.append_slice(&[attr_flags, attr_typecode]);
+            let _ = self.target.append_slice(&(attr_len as u16).to_be_bytes());
+            let _ = self.target.append_slice(&asp);
+
+            total_pa_len += 2 + 2 + attr_len;
+        }
+
+
+        // XXX the next_hop is either a (conventional, for v4/unicast) path
+        // attribute, or, it is part of MP_REACH_NLRI.
+        // Should/must v4/unicast always go in MP_REACH_NLRI when both peers
+        // sent such capability though?
+        if let Some(next_hop) = acs.next_hop.into_opt() {
+            match next_hop {
+                NextHop::Ipv4(v4addr) => {
+                    let attr_flags = 0b0100_0000;
+                    let attr_typecode = PathAttributeType::NextHop.into();
+                    let attr_len = 4_u8; 
+
+                    let _ = self.target.append_slice(
+                        &[attr_flags, attr_typecode, attr_len]
                     );
-                });
+                    let _ = self.target.append_slice(&v4addr.octets());
 
-                MpUnreachNlri(builder).into()
+                    total_pa_len += 2 + 1 + usize::from(attr_len);
                 }
-            );
-
-            check(
-                vec![
-                    0xc0, 0x10, 0x08, 0x00, 0x02, 0xfc, 0x85, 0x00,
-                    0x00, 0xcf, 0x08
-                ],
-                ExtendedCommunities(ExtendedCommunitiesList::new(vec![
-                    "rt:64645:53000".parse().unwrap()
-                ])).into()
-            );
-
-            check(
-                vec![0xc0, 0x11, 10,
-                0x02, 0x02, // SEQUENCE of length 2
-                0x00, 0x00, 0x00, 100,
-                0x00, 0x00, 0x00, 200,
-                ],
-                PA::As4Path(As4Path(HopPath::from(vec![
-                    Asn::from_u32(100),
-                    Asn::from_u32(200)]
-                )))
-            );
-
-            check(
-                vec![
-                    0xc0, 0x12, 0x08, 0x00, 0x00, 0x04, 0xd2,
-                    10, 0, 0, 99
-                ],
-                As4Aggregator(AggregatorInfo::new(
-                    Asn::from_u32(1234),
-                    "10.0.0.99".parse().unwrap()
-                )).into()
-            );
-
-            check(
-                vec![0xc0, 0x14, 0x04, 1, 2, 3, 4],
-                Connector("1.2.3.4".parse().unwrap()).into()
-            );
-
-            check(
-                vec![0xc0, 0x15, 0x05, 0x14, 0x00, 0x00, 0x04, 0xd2],
-                AsPathLimit(
-                    AsPathLimitInfo::new(20, Asn::from_u32(1234))
-                ).into()
-            );
-
-            //TODO 22 PmsiTunnel
-            //TODO 25 Ipv6ExtendedCommunities
-
-            check(
-                vec![
-                    0xc0, 0x20, 0x3c, 0x00, 0x00, 0x20, 0x5b, 0x00,
-                    0x00, 0x00, 0x06, 0x00, 0x00, 0x00, 0x0f, 0x00,
-                    0x00, 0xe2, 0x0a, 0x00, 0x00, 0x00, 0x64, 0x00,
-                    0x00, 0x0b, 0x62, 0x00, 0x00, 0xe2, 0x0a, 0x00,
-                    0x00, 0x00, 0x65, 0x00, 0x00, 0x00, 0x64, 0x00,
-                    0x00, 0xe2, 0x0a, 0x00, 0x00, 0x00, 0x67, 0x00,
-                    0x00, 0x00, 0x01, 0x00, 0x00, 0xe2, 0x0a, 0x00,
-                    0x00, 0x00, 0x68, 0x00, 0x00, 0x00, 0x1f
-                ],
-                LargeCommunities(LargeCommunitiesList::new(
-                    vec![
-                        "AS8283:6:15".parse().unwrap(),
-                        "AS57866:100:2914".parse().unwrap(),
-                        "AS57866:101:100".parse().unwrap(),
-                        "AS57866:103:1".parse().unwrap(),
-                        "AS57866:104:31".parse().unwrap(),
-                    ]
-                )).into()
-            );
-
-            check(
-                vec![0xc0, 0x23, 0x04, 0x00, 0x00, 0x04, 0xd2],
-                Otc(Asn::from_u32(1234)).into()
-            );
-
-            // TODO AttrSet
-            // TODO Reserved?
-
-            // UnimplementedPathAttribute
-            check(
-                vec![0xc0, 254, 0x04, 0x01, 0x02, 0x03, 0x04],
-                UnimplementedPathAttribute::new(
-                    Flags::OPT_TRANS.into(),
-                    254,
-                    vec![0x01, 0x02, 0x03, 0x04]
-                ).into()
-            );
-
-
-
-        }
-
-        #[test]
-        fn iter_and_find() {
-            let raw = vec![
-                0x40, 0x01, 0x01, 0x00, // ORIGIN
-                0x40, 0x03, 0x04, 0x01, 0x02, 0x03, 0x04, // NEXTHOP
-                0x80, 0x04, 0x04, 0x00, 0x00, 0x00, 0xff // MED 
-            ];
-            let pas = PathAttributes::new(
-                Parser::from_ref(&raw), SessionConfig::modern()
-            );
-            //for _ in 0..4 {
-            //    let pa = pas.next();
-            //    println!("{pa:?}");
-            //}
-
-            assert!(pas.get(PathAttributeType::Origin).is_some());
-            assert!(pas.get(PathAttributeType::AsPath).is_none());
-            assert!(pas.get(PathAttributeType::MultiExitDisc).is_some());
-            assert!(pas.get(PathAttributeType::NextHop).is_some());
-        }
-
-        #[test]
-        fn unimplemented_path_attributes() {
-            let raw = vec![
-                0xc0, 254, 0x04, 0x01, 0x02, 0x03, 0x04
-            ];
-            let mut parser = Parser::from_ref(&raw);
-            let sc = SessionConfig::modern();
-            let pa = WireformatPathAttribute::parse(&mut parser, sc);
-
-            if let Ok(WireformatPathAttribute::Unimplemented(u)) = pa {
-                assert_eq!(u.typecode(), 254);
-                assert!(u.flags().is_optional());
-                assert!(u.flags().is_transitive());
-                assert_eq!(u.value(), &[0x01, 0x02, 0x03, 0x04]);
-            } else {
-                panic!("fail");
+                _ => todo!() // this is MP_REACH_NLRI territory
             }
-
         }
 
-        #[test]
-        fn parse_unexpected_two_octet_asn() {
-            let raw = vec![
-                0xc0, 0x07, 0x06, 0x00, 0x65, 0xc6,
-                0x33, 0x64, 0x01
-            ];
-            let mut parser = Parser::from_ref(&raw);
-            let pa = WireformatPathAttribute::parse(
-                    &mut parser, SessionConfig::modern()
-            ).unwrap();
-            assert!(matches!(pa, WireformatPathAttribute::Invalid(_,_,_)));
-        }
 
-        /*
-        #[test]
-        fn deref_mut() {
-            // AS_PATH: AS_SEQUENCE(AS100, AS200)
-            let raw = vec![0x40, 0x02, 10,
-                0x02, 0x02, // SEQUENCE of length 2
-                0x00, 0x00, 0x00, 100,
-                0x00, 0x00, 0x00, 200,
-            ];
+        if let Some(comms) = acs.standard_communities.into_opt() {
+            let attr_flags = 0b0100_0000;
+            let attr_typecode = PathAttributeType::Communities.into();
+            let attr_len = match u8::try_from(4 * comms.len()) {
+                Ok(n) => n,
+                Err(..) => {
+                    return Err(ComposeError::AttributeTooLarge(
+                        PathAttributeType::Communities,
+                        4 * comms.len()
+                    ));
+                }
+            };
 
-            let pa = WireformatPathAttribute::parse(
-                &mut Parser::from_ref(&raw), SessionConfig::modern()
-            ).unwrap();
-            let mut owned = pa.to_owned().unwrap();
-            if let PathAttribute::AsPath(ref mut asp) = owned  {
-                assert_eq!(format!("{}", **asp), "AS100 AS200");
-                asp.prepend(Asn::from_u32(50));
-                assert_eq!(format!("{}", **asp), "AS50 AS100 AS200");
-                assert_eq!(3, asp.hop_count());
+            let _ = self.target.append_slice(
+                &[attr_flags, attr_typecode, attr_len]
+            );
+
+            for c in comms {
+                let _ = self.target.append_slice(&c.to_raw());
             }
-
-            let mut composed = Vec::new();
-            owned.compose(&mut composed).unwrap();
-            assert!(composed != raw);
+            total_pa_len += 2 + 1 + usize::from(attr_len);
         }
-        */
 
-        #[test]
-        fn as_ref_as_mut() {
-            // AS_PATH: AS_SEQUENCE(AS100, AS200)
-            let raw = vec![0x40, 0x02, 10,
-                0x02, 0x02, // SEQUENCE of length 2
-                0x00, 0x00, 0x00, 100,
-                0x00, 0x00, 0x00, 200,
-            ];
 
-            let pa = WireformatPathAttribute::parse(
-                &mut Parser::from_ref(&raw), SessionConfig::modern()
-            ).unwrap();
-            let mut owned = pa.to_owned().unwrap();
-            if let PathAttribute::AsPath(ref mut asp) = owned  {
-                assert_eq!(format!("{}", asp.as_ref()), "AS100 AS200");
-                asp.as_mut().prepend(Asn::from_u32(50));
-                assert_eq!(format!("{}", asp.as_ref()), "AS50 AS100 AS200");
-                assert_eq!(3, asp.as_ref().hop_count());
+        if u16::try_from(total_pa_len).is_err() {
+            return Err(ComposeError::AttributesTooLarge(total_pa_len));
+        }
+
+        // update total path attribute len:
+        self.target.as_mut()[21+withdraw_len..21+withdraw_len+2]
+            .copy_from_slice(&(total_pa_len as u16).to_be_bytes());
+
+
+        // NLRI
+        // TODO this all needs to be a lot more sophisticated:
+        //  - prefixes can not occur in both withdrawals and nlris, so check
+        //  for that;
+        //  - non v4/unicast NLRI should go in MP_REACH_NLRI, not here (at the
+        //  end of the PDU);
+        //  - we should be able to put multiple NLRI in one UPDATE, though
+        //  currently the AttrChangeSet only holds one;
+        //  - probably more
+        
+        let mut nlri_len = 0;
+
+        if let Some(nlri) = acs.nlri.into_opt() {
+            match nlri {
+                Nlri::Unicast(b) => {
+                    if let Some(id) = b.path_id() {
+                        let _ = self.target.append_slice(&id.to_raw());
+                        nlri_len += 4;
+                    }
+                    match b.prefix().addr_and_len() {
+                        (std::net::IpAddr::V4(addr), len) => {
+                            let _ = self.target.append_slice(&[len]);
+                            let len_bytes = (usize::from(len)-1) / 8 + 1;
+                            let _ = self.target.append_slice(
+                                &addr.octets()[0..len_bytes]
+                            );
+                            nlri_len += 1 + len_bytes;
+                        }
+                        _ => todo!()
+                    }
+                }
+                _ => todo!()
             }
+        }
 
-            let mut composed = Vec::new();
-            owned.compose(&mut composed).unwrap();
-            assert!(composed != raw);
+        // update pdu len
+        let msg_len = 19 
+            + 2 + withdraw_len 
+            + 2 + total_pa_len
+            + nlri_len
+        ;
+
+        if msg_len > 4096 {
+            // TODO handle Extended Messages (max pdu size 65535)
+            return Err(ComposeError::PduTooLarge(msg_len));
+        }
+
+        //if u16::try_from(msg_len).is_err() {
+        //    return Err(ComposeError::PduTooLarge(msg_len));
+        //}
+
+        self.target.as_mut()[16..=17].copy_from_slice(
+            &(msg_len as u16).to_be_bytes()
+        );
+
+        Ok(self.target)
+    }
+}
+
+impl<Target> UpdateBuilder<Target>
+where
+    Target: OctetsBuilder + FreezeBuilder + AsMut<[u8]>,
+    <Target as FreezeBuilder>::Octets: Octets,
+{
+    pub fn into_message(self) ->
+        Result<UpdateMessage<<Target as FreezeBuilder>::Octets>, ComposeError>
+    where
+        Target: FreezeBuilder,
+        <Target as FreezeBuilder>::Octets: Octets,
+    {
+        self.is_valid()?;
+        Ok(UpdateMessage::from_octets(
+            self.finish().map_err(|_| ShortBuf)?, SessionConfig::modern()
+        )?)
+    }
+
+    // Check whether the combination of NLRI and attributes would produce a
+    // valid UPDATE pdu.
+    fn is_valid(&self) -> Result<(), ComposeError> {
+        // If we have builders for MP_(UN)REACH_NLRI, they should carry
+        // prefixes.
+        if let Some(ref builder) = self.mp_reach_nlri_builder {
+            if builder.is_empty() {
+                return Err(ComposeError::EmptyMpReachNlri);
+            }
+        }
+        if let Some(ref builder) = self.mp_unreach_nlri_builder {
+            if builder.is_empty() {
+                return Err(ComposeError::EmptyMpUnreachNlri);
+            }
+        }
+
+        Ok(())
+    }
+
+    fn finish(mut self)
+        -> Result<<Target as FreezeBuilder>::Octets, Target::AppendError>
+    {
+        let mut header = Header::for_slice_mut(self.target.as_mut());
+        header.set_length(u16::try_from(self.total_pdu_len).unwrap());
+
+        // `withdrawals_len` is checked to be <= 4096 or <= 65535
+        // so it will always fit in a u16.
+        self.target.append_slice(
+            &u16::try_from(self.withdrawals_len).unwrap().to_be_bytes()
+        )?;
+
+        for w in &self.withdrawals {
+            match w {
+                Nlri::Unicast(b) => {
+                    if b.is_v4() {
+                        b.compose(&mut self.target)?;
+                    } else {
+                        // Other withdrawals should not go here, but in
+                        // MP_UNREACH_NLRI. Handling these ones below in the
+                        // path attributes.
+                    }
+                },
+                _ => todo!(),
+            }
         }
 
 
+
+        // XXX we can do these unwraps because of the checks in the add/append
+        // methods
+
+        // `attributes_len` is checked to be <= 4096 or <= 65535
+        // so it will always fit in a u16.
+        let _ = self.target.append_slice(
+            &u16::try_from(self.attributes_len).unwrap().to_be_bytes()
+        );
+
+        // TODO write all path attributes, if any, in order of typecode
+
+        if let Some(origin) = self.origin {
+            origin.compose(&mut self.target)?
+        }
+
+        if let Some(aspath) = self.aspath {
+            aspath.compose(&mut self.target)?
+        }
+
+        if let Some(nexthop) = self.nexthop {
+            nexthop.compose(&mut self.target)?
+        }
+
+        if let Some(med) = self.multi_exit_disc {
+            med.compose(&mut self.target)?
+        }
+
+        if let Some(local_pref) = self.local_pref {
+            local_pref.compose(&mut self.target)?
+        }
+
+        if let Some(builder) = self.standard_communities_builder {
+            builder.compose(&mut self.target)?
+        }
+
+        if let Some(builder) = self.mp_reach_nlri_builder {
+            builder.compose(&mut self.target)?
+        }
+
+        if let Some(builder) = self.mp_unreach_nlri_builder {
+            builder.compose(&mut self.target)?
+        }
+
+        // XXX Here, in the conventional NLRI field at the end of the PDU, we
+        // write IPv4 Unicast announcements. But what if we have agreed to do
+        // MP for v4/unicast, should these announcements go in the
+        // MP_REACH_NLRI attribute then instead?
+        for a in &self.announcements {
+            match a {
+                Nlri::Unicast(b) => {
+                    if b.is_v4() {
+                        b.compose(&mut self.target)?;
+                    } else {
+                        // Other announcements should not go here, but in
+                        // MP_REACH_NLRI, handled before in the path
+                        // attributes.
+                    }
+                },
+                _ => todo!(),
+            }
+        }
+
+        Ok(self.target.freeze())
+    }
+}
+
+impl UpdateBuilder<Vec<u8>> {
+    pub fn new_vec() -> Self {
+        Self::from_target(Vec::with_capacity(23)).unwrap()
+    }
+}
+
+impl UpdateBuilder<BytesMut> {
+    pub fn new_bytes() -> Self {
+        Self::from_target(BytesMut::new()).unwrap()
     }
 }
 
@@ -2350,6 +1169,9 @@ impl MpUnreachNlriBuilder {
     }
 }
 
+//------------ StandardCommunitiesBuilder ------------------------------------
+//
+//
 
 #[derive(Debug)]
 pub(crate) struct StandardCommunitiesBuilder {
@@ -2402,5 +1224,79 @@ impl StandardCommunitiesBuilder {
             target.append_slice(&c.to_raw())?;
         }
         Ok(())
+    }
+}
+
+//------------ Errors --------------------------------------------------------
+
+#[derive(Debug)]
+pub enum ComposeError{
+    /// Exceeded maximum PDU size, data field carries the violating length.
+    PduTooLarge(usize),
+
+    // TODO proper docstrings, first see how/if we actually use these.
+    AttributeTooLarge(new_pas::PathAttributeType, usize),
+    AttributesTooLarge(usize),
+    IllegalCombination,
+    EmptyMpReachNlri,
+    EmptyMpUnreachNlri,
+    WrongAddressType,
+
+    InvalidAttribute,
+
+    /// Variant for `octseq::builder::ShortBuf`
+    ShortBuf,
+    /// Wrapper for util::parser::ParseError, used in `fn into_message`
+    ParseError(ParseError)
+}
+
+impl From<ShortBuf> for ComposeError {
+    fn from(_: ShortBuf) -> ComposeError {
+        ComposeError::ShortBuf
+    }
+}
+impl From<ParseError> for ComposeError {
+    fn from(pe: ParseError) -> ComposeError {
+        ComposeError::ParseError(pe)
+    }
+}
+
+
+impl std::error::Error for ComposeError { }
+impl fmt::Display for ComposeError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            ComposeError::PduTooLarge(n) => {
+                write!(f, "oversized PDU: {n} bytes")
+            }
+            ComposeError::AttributeTooLarge(attr, n) => {
+                write!(f, "oversized attribute {attr}: {n} bytes")
+            }
+            ComposeError::AttributesTooLarge(n) => {
+                write!(f, "total path attributes too large: {n} bytes")
+            }
+            ComposeError::IllegalCombination => {
+                write!(f, "illegal combination of prefixes/attributes")
+            }
+            ComposeError::EmptyMpReachNlri => {
+                write!(f, "missing NLRI in MP_REACH_NLRI")
+            }
+            ComposeError::EmptyMpUnreachNlri => {
+                write!(f, "missing NLRI in MP_UNREACH_NLRI")
+            }
+            ComposeError::WrongAddressType => {
+                write!(f, "wrong address type")
+            }
+            ComposeError::InvalidAttribute => {
+                write!(f, "invalid attribute")
+            }
+            ComposeError::ShortBuf => {
+                ShortBuf.fmt(f)
+            }
+            ComposeError::ParseError(pe) => {
+                write!(f, "parse error in builder: {}", pe)
+            }
+
+        }
     }
 }
